Http/https implementation.  InstanceProps implementation.  Screen to edit InstanceProps.  New var to control whether scheduled tasks run on this instance.

+++ source/WEB-INF/classes/com/dneero/scheduledjobs/DeleteOldPersistentlogins.java
@@ -12,6 +12,7 @@ import com.dneero.dao.hibernate.HibernateUtil;
 import com.dneero.util.GeneralException;
 import com.dneero.util.Time;
 import com.dneero.session.PersistentLogin;
+import com.dneero.systemprops.InstanceProperties;
 
 import java.util.List;
 import java.util.Date;
@@ -28,19 +29,20 @@ public class DeleteOldPersistentlogins implements Job {
     Logger logger = Logger.getLogger(this.getClass().getName());
 
     public void execute(JobExecutionContext jobExecutionContext) throws JobExecutionException {
-        logger.debug("execute() DeleteOldPersistentlogins called");
+        if (InstanceProperties.getRunScheduledTasksOnThisInstance()){
+            logger.debug("execute() DeleteOldPersistentlogins called");
 
 
-        List<Userpersistentlogin> userpersistentlogins = HibernateUtil.getSession().createQuery("from Userpersistentlogin").list();
-        for (Iterator iterator = userpersistentlogins.iterator(); iterator.hasNext();) {
-            Userpersistentlogin userpersistentlogin = (Userpersistentlogin) iterator.next();
-            Date longtimeago = Time.xDaysAgoEnd(Calendar.getInstance(), PersistentLogin.daysToKeepPersistentRecordWithoutLogin).getTime();
-            if (userpersistentlogin.getLastusedtologin().before(longtimeago)){
-               try{userpersistentlogin.delete();}catch(Exception ex){logger.error(ex);};
+            List<Userpersistentlogin> userpersistentlogins = HibernateUtil.getSession().createQuery("from Userpersistentlogin").list();
+            for (Iterator iterator = userpersistentlogins.iterator(); iterator.hasNext();) {
+                Userpersistentlogin userpersistentlogin = (Userpersistentlogin) iterator.next();
+                Date longtimeago = Time.xDaysAgoEnd(Calendar.getInstance(), PersistentLogin.daysToKeepPersistentRecordWithoutLogin).getTime();
+                if (userpersistentlogin.getLastusedtologin().before(longtimeago)){
+                   try{userpersistentlogin.delete();}catch(Exception ex){logger.error(ex);};
+                }
             }
+
         }
-
-
 
 
     }
